Refactoring Id of Checkin entity to Long

+++ app/src/main/java/com/alienonwork/crossfitcheckin/services/boundServices/ScheduleService.java
@@ -64,7 +64,7 @@ public class ScheduleService extends LifecycleService {
 
             if (workInfo.getTags().contains(PostCheckinWorker.TAG)) {
                 Data data = workInfo.getOutputData();
-                Integer checkinId = data.getInt(PostCheckinWorker.PARAM_CHECKIN_ID, 0);
+                Long checkinId = data.getLong(PostCheckinWorker.PARAM_CHECKIN_ID, 0);
 
                 if (workInfo.getState().isFinished() && workInfo.getState() == WorkInfo.State.SUCCEEDED) {
                     // Notifica o usuário que o checkin foi feito
@@ -130,7 +130,7 @@ public class ScheduleService extends LifecycleService {
         super.onStartCommand(intent, flags, startId);
 
         if (intent.getBooleanExtra(ScheduleService.EXTRA_SCHEDULE_CHECKIN, false)) {
-            postCheckin(intent.getIntExtra(PostCheckinWorker.PARAM_CHECKIN_ID, 0), TAG_SETUP_SCHEDULE);
+            postCheckin(intent.getLongExtra(PostCheckinWorker.PARAM_CHECKIN_ID, 0), TAG_SETUP_SCHEDULE);
         }
 
         if (intent.getBooleanExtra(ScheduleService.EXTRA_RESCHEDULE_CHECKIN, false)) {
@@ -276,14 +276,14 @@ public class ScheduleService extends LifecycleService {
         workManager.getWorkInfoByIdLiveData(getCheckin.getId()).observe(this, workerObserver);
     }
 
-    private void postCheckin(int id, String tag) {
+    private void postCheckin(Long id, String tag) {
         postCheckin(id, tag,false);
     }
 
-    private void postCheckin(int id, String tag, Boolean networkRequired) {
+    private void postCheckin(Long id, String tag, Boolean networkRequired) {
         Data.Builder builder = new Data.Builder();
 
-        builder.putInt(PostCheckinWorker.PARAM_CHECKIN_ID, id);
+        builder.putLong(PostCheckinWorker.PARAM_CHECKIN_ID, id);
 
         Data data = builder.build();
